// packages/scraper/common/types.ts
/**
 * Shared types for scraper package.
 *
 * These string literal unions correspond to the enum values defined in
 * packages/viewer/constants/enums.ts (ReservationDivision, ReservationStatus).
 */

export type Division =
  | "RESERVATION_DIVISION_INVALID"
  | "RESERVATION_DIVISION_MORNING"
  | "RESERVATION_DIVISION_MORNING_ONE"
  | "RESERVATION_DIVISION_MORNING_TWO"
  | "RESERVATION_DIVISION_AFTERNOON"
  | "RESERVATION_DIVISION_AFTERNOON_ONE"
  | "RESERVATION_DIVISION_AFTERNOON_TWO"
  | "RESERVATION_DIVISION_EVENING"
  | "RESERVATION_DIVISION_EVENING_ONE"
  | "RESERVATION_DIVISION_EVENING_TWO"
  | "RESERVATION_DIVISION_DIVISION_1"
  | "RESERVATION_DIVISION_DIVISION_2"
  | "RESERVATION_DIVISION_DIVISION_3"
  | "RESERVATION_DIVISION_DIVISION_4"
  | "RESERVATION_DIVISION_DIVISION_5"
  | "RESERVATION_DIVISION_DIVISION_6"
  | "RESERVATION_DIVISION_DIVISION_7"
  | "RESERVATION_DIVISION_DIVISION_8"
  | "RESERVATION_DIVISION_DIVISION_9"
  | "RESERVATION_DIVISION_DIVISION_10"
  | "RESERVATION_DIVISION_DIVISION_11"
  | "RESERVATION_DIVISION_DIVISION_12"
  | "RESERVATION_DIVISION_DIVISION_13"
  | "RESERVATION_DIVISION_DIVISION_14"
  | "RESERVATION_DIVISION_DIVISION_15"
  | "RESERVATION_DIVISION_DIVISION_16"
  | "RESERVATION_DIVISION_DIVISION_17"
  | "RESERVATION_DIVISION_DIVISION_18"
  | "RESERVATION_DIVISION_DIVISION_19"
  | "RESERVATION_DIVISION_DIVISION_20"
  | "RESERVATION_DIVISION_DIVISION_21"
  | "RESERVATION_DIVISION_DIVISION_22"
  | "RESERVATION_DIVISION_DIVISION_23"
  | "RESERVATION_DIVISION_DIVISION_24"
  | "RESERVATION_DIVISION_DIVISION_25"
  | "RESERVATION_DIVISION_DIVISION_26"
  | "RESERVATION_DIVISION_DIVISION_27"
  | "RESERVATION_DIVISION_DIVISION_28"
  | "RESERVATION_DIVISION_DIVISION_29"
  | "RESERVATION_DIVISION_DIVISION_30";

export type Status =
  | "RESERVATION_STATUS_INVALID"
  | "RESERVATION_STATUS_VACANT"
  | "RESERVATION_STATUS_STATUS_1"
  | "RESERVATION_STATUS_STATUS_2"
  | "RESERVATION_STATUS_STATUS_3"
  | "RESERVATION_STATUS_STATUS_4"
  | "RESERVATION_STATUS_STATUS_5"
  | "RESERVATION_STATUS_STATUS_6"
  | "RESERVATION_STATUS_STATUS_7"
  | "RESERVATION_STATUS_STATUS_8"
  | "RESERVATION_STATUS_STATUS_9"
  | "RESERVATION_STATUS_STATUS_10"
  | "RESERVATION_STATUS_STATUS_11"
  | "RESERVATION_STATUS_STATUS_12"
  | "RESERVATION_STATUS_STATUS_13"
  | "RESERVATION_STATUS_STATUS_14"
  | "RESERVATION_STATUS_STATUS_15"
  | "RESERVATION_STATUS_STATUS_16"
  | "RESERVATION_STATUS_STATUS_17"
  | "RESERVATION_STATUS_STATUS_18"
  | "RESERVATION_STATUS_STATUS_19"
  | "RESERVATION_STATUS_STATUS_20";

export type Reservation = { [K in Division]?: Status };

export type TransformOutput = {
  room_name: string;
  date: string;
  reservation: Reservation;
}[];